Combine examples for different transports.
Servers now listen on both transports and clients choose the transport to use
via a command line argument.
