+++ examples/thrift/keyvalue/client/.gitignore
