--- examples/thrift/keyvalue/http/server/.gitignore
@@ -1,1 +1,0 @@
-server
