+++ examples/thrift/keyvalue/server/.gitignore
