--- examples/thrift/keyvalue/tchannel/client/.gitignore
@@ -1,1 +1,0 @@
-client
