--- examples/thrift/keyvalue/tchannel/server/.gitignore
@@ -1,1 +1,0 @@
-server
